// apps/admin-web/src/api/model/baseResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Admin API
 * The Admin API description
 * OpenAPI spec version: 1.0
 */
import type { ErrorDetail } from './errorDetail';

export interface BaseResponse {
  /** 성공 여부 */
  success: boolean;
  /** 응답 코드 */
  code: string;
  /**
   * 응답 메시지
   * @nullable
   */
  message?: string | null;
  /** 응답 생성 시간 */
  timestamp: string;
  /** 요청 ID */
  requestId: string;
  /**
   * 에러 목록
   * @nullable
   */
  errors?: ErrorDetail[] | null;
}